requirements.txt
python-telegram-bot>=20.0
fastapi>=0.100.0
uvicorn>=0.23.0
openai>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.10
qdrant-client>=1.3.0
openrouter>=0.3.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
httpx>=0.25.0
treesitter-chunker>=2.0.0
pathspec>=1.0.0
